correct buttot modal and js
