+++ src/index.js
@@ -1,9 +1,4 @@
-
-import './js/customers-reviews.js'
-import './js/test'
-import './js/mobile-menu'
-import './js/modal-test'
-import './js/modal'
+import './js/swiper';
+import './js/mobile-menu';
+import './js/main-modal';
 // import './js/gallery.js'
-
-
