+++ src/partials/contacts.html
@@ -1,6 +1,6 @@
 <section class="section contacts__section">
     <div class="container contacts__container">
-        <h2 class="visually-hidden">Our contacts</h2>
+        <h2 id="contact" class="visually-hidden">Our contacts</h2>
         <ul class="contacts__list">
             <li class="contacts__card">
                 <p class="contacts__subtitle">cafe</p>
