+++ src/partials/header.html
@@ -1,83 +1,77 @@
 <header class="header">
-    <section class="section head">
-        <div class="container nav">
-            <img
-                class="company__logo"
-                srcset="
-                    ./images/hero-header/logo.png             48w,
-                    ./images/hero-header/logo@2x.png          96w,
-                    ./images/hero-header/logo-desktop.png     60w,
-                    ./images/hero-header/logo-desktop@2x.png 120w
-                "
-                sizes="(min-width: 1200px) 60px, (min-width: 320px) 48px, 100vw"
-                src="./images/hero-header/logo.png"
-                alt="ice cream logo"
-                loading="lazy"
-            />
-            <button
-                type="button"
-                class="button__menu"
-                data-menu-btn
-                aria-controls="mobile__menu-body"
+    <div class="container nav">
+        <h2 class="visually-hidden">Our website</h2>
+        <img
+            class="company__logo"
+            srcset="
+                ./images/hero-header/logo.png             48w,
+                ./images/hero-header/logo@2x.png          96w,
+                ./images/hero-header/logo-desktop.png     60w,
+                ./images/hero-header/logo-desktop@2x.png 120w
+            "
+            sizes="(min-width: 1200px) 60px, (min-width: 320px) 48px, 100vw"
+            src="./images/hero-header/logo.png"
+            alt="ice cream logo"
+            loading="lazy"
+        />
+        <button type="button" class="button__menu" data-menu-btn>
+            <svg
+                class="button__menu-icons"
+                width="32"
+                height="32"
+                aria-label="Вызов мобильного меню"
             >
-                <svg
-                    class="button__menu-icons"
-                    width="32"
-                    height="32"
-                    aria-label="Вызов мобильного меню"
-                >
-                    <use
-                        class="button__menu-open"
-                        href="./images/icons.svg#icon-open-mobile-menu"
-                    ></use>
-                    <use
-                        class="button__menu-close"
-                        href="./images/icons.svg#icon-close-mobile-menu"
-                    ></use>
+                <use
+                    class="button__menu-open"
+                    href="./images/icons.svg#icon-open-mobile-menu"
+                ></use>
+                <use
+                    class="button__menu-close"
+                    href="./images/icons.svg#icon-close-mobile-menu"
+                ></use>
+            </svg>
+        </button>
+        <div class="nav__buy-body">
+            <button data-btn="buy" type="button" class="nav__buy-button">
+                Buy now
+                <svg class="nav__buy-icon" width="12" height="12">
+                    <use href="./images/icons.svg#about__icon"></use>
                 </svg>
             </button>
-            <div class="nav__buy-body">
-                <button data-btn="buy" type="button" class="nav__buy-button">
+        </div>
+
+        <div class="nav__body" data-mobile-menu>
+            <nav>
+                <ul class="nav__list">
+                    <li class="nav__item">
+                        <a href="" class="nav__link">Home</a>
+                    </li>
+                    <li class="nav__item">
+                        <a href="#about" class="nav__link">How it’s made?</a>
+                    </li>
+                    <li class="nav__item">
+                        <a href="#products" class="nav__link">Our products</a>
+                    </li>
+                    <li class="nav__item">
+                        <a href="#contact" class="nav__link">Contact</a>
+                    </li>
+                </ul>
+            </nav>
+            <div class="nav__button-body">
+                <button data-btn="buy" data-modal-open type="button" class="nav__button-buy">
                     Buy now
-                    <svg data-btn="buy" class="nav__buy-icon" width="12" height="12">
-                        <use data-btn="buy" href="./images/icons.svg#about__icon"></use>
+                    <svg class="nav__icon-buy" width="12" height="12">
+                        <use href="./images/icons.svg#about__icon"></use>
                     </svg>
                 </button>
             </div>
-
-            <div class="nav__body" data-mobile-menu>
-                <nav>
-                    <ul class="nav__list">
-                        <li class="nav__item">
-                            <a href="" class="nav__link">Home</a>
-                        </li>
-                        <li class="nav__item">
-                            <a href="#about" class="nav__link">How it’s made?</a>
-                        </li>
-                        <li class="nav__item">
-                            <a href="#products" class="nav__link">Our products</a>
-                        </li>
-                        <li class="nav__item">
-                            <a href="" class="nav__link">Contact</a>
-                        </li>
-                    </ul>
-                </nav>
-                <div class="nav__button-body">
-                    <button data-btn="buy" data-modal-open type="button" class="nav__button-buy">
-                        Buy now
-                        <svg data-btn="buy" class="nav__icon-buy" width="12" height="12">
-                            <use data-btn="buy" href="./images/icons.svg#about__icon"></use>
-                        </svg>
-                    </button>
-                </div>
-            </div>
         </div>
-    </section>
+    </div>
     <div class="backdrop is-hidden" id="buy">
         <div class="modal">
-            <button class="modal__close-button" type="button" data-btn="close">
+            <button class="modal__close-button" type="button">
                 <svg class="modal__close-icon" width="20" height="20" data-btn="close">
-                    <use href="./images/icons.svg#icon-close-modal" data-btn="close"></use>
+                    <use href="./images/icons.svg#icon-close-modal"></use>
                 </svg>
             </button>
             <form class="modal__form">
