+++ src/sass/layout/_header.scss
@@ -203,7 +203,7 @@
         fill: $text-color-light;
     }
 
-    @media screen and (min-width: 767px) and (max-width: 1199px) {
+    @media screen and (min-width: 768px) and (max-width: 1199px) {
         display: block;
         margin-left: 38px;
     }
